# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Gateway API
  version: 0.1.0-draft
  description: |
    Платформенный каталог инструментов и установки в workspace.
    x-erd: cursor-context/docs/erd-and-bounded-contexts.md#4-bc-mcp-registry--proxy-mcp-gateway
servers:
  - url: http://localhost:8081/v1/mcp
    description: локально
tags:
  - name: catalog
  - name: installations
  - name: internal
security:
  - bearerAuth: []
paths:
  /catalog:
    get:
      tags: [catalog]
      operationId: mcpCatalogList
      summary: Каталог MCP (платформа)
      security: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogListResponse'
  /catalog/{connectorKey}:
    get:
      tags: [catalog]
      operationId: mcpCatalogGet
      parameters:
        - $ref: '#/components/parameters/ConnectorKey'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogTool'
  /workspaces/{workspaceId}/installations:
    get:
      tags: [installations]
      operationId: mcpInstallationsList
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationListResponse'
    post:
      tags: [installations]
      operationId: mcpInstallationsCreate
      summary: Установить из каталога (credentials + policy pack apply)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Installation'
        '502':
          description: policy_apply_failed — откат установки
  /workspaces/{workspaceId}/installations/{installationId}:
    patch:
      tags: [installations]
      operationId: mcpInstallationsPatch
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/InstallationId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Installation'
    delete:
      tags: [installations]
      operationId: mcpInstallationsDelete
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/InstallationId'
      responses:
        '204':
          description: OK
  /workspaces/{workspaceId}/installations/{installationId}/health:
    post:
      tags: [installations]
      operationId: mcpInstallationsHealth
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/InstallationId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResult'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstallationId:
      name: installationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ConnectorKey:
      name: connectorKey
      in: path
      required: true
      schema:
        type: string
        examples: [notion]
  schemas:
    CatalogTool:
      type: object
      required: [connectorKey, displayName, connectionFormSchema]
      properties:
        connectorKey:
          type: string
        displayName:
          type: string
        connectionFormSchema:
          type: object
          description: JSON Schema для формы установки
        policyPackVersion:
          type: string
    CatalogListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogTool'
    Installation:
      type: object
      required: [id, workspaceId, connectorKey, displayLabel, status]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        connectorKey:
          type: string
        displayLabel:
          type: string
        status:
          type: string
          enum: [active, error, revoked]
    InstallationCreateRequest:
      type: object
      required: [connectorKey, form]
      properties:
        connectorKey:
          type: string
        displayLabel:
          type: string
        form:
          type: object
          additionalProperties: true
          description: Поля по connection_form_schema
    InstallationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Installation'
    HealthCheckResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string